refactor: Use default protobuf in Rust
Instead of using copy of protobuf files in local Rust project. Generate
at compile time.
