+++ rust/remotivelabs-broker/build.rs
@@ -1,17 +1,17 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     tonic_build::configure()
         .build_server(false)
-        .out_dir("src/beamy_api")
+        .out_dir("src/remotive_api")
         .compile(
             &[
-                "protos/common.proto",
-                "protos/diagnostics_api.proto",
-                "protos/functional_api.proto",
-                "protos/network_api.proto",
-                "protos/system_api.proto",
-                "protos/traffic_api.proto",
+                "common.proto",
+                "diagnostics_api.proto",
+                "functional_api.proto",
+                "network_api.proto",
+                "system_api.proto",
+                "traffic_api.proto",
             ],
-            &["protos"],
+            &["../../protos"],
         )?;
 
     Ok(())
