+++ rust/remotivelabs-broker/examples/client/build.rs
@@ -4,14 +4,14 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         .out_dir("src/beamy_api")
         .compile(
             &[
-                "protos/common.proto",
-                "protos/diagnostics_api.proto",
-                "protos/functional_api.proto",
-                "protos/network_api.proto",
-                "protos/system_api.proto",
-                "protos/traffic_api.proto",
+                "common.proto",
+                "diagnostics_api.proto",
+                "functional_api.proto",
+                "network_api.proto",
+                "system_api.proto",
+                "traffic_api.proto",
             ],
-            &["protos"],
+            &["../../../protos"],
         )?;
     Ok(())
 }
