+++ rust/remotivelabs-broker/examples/pubsub/Cargo.toml
@@ -18,4 +18,4 @@ tokio = { version = "1.14.0", features = ["rt-multi-thread", "time", "fs", "macr
 tonic = "0.6.2"
 
 # local lib
-lib-helper = { path = "../helper" }
+lib-helper = { path = "../" }
