--- rust/remotivelabs-broker/examples/pubsub/protos/common.proto
@@ -1,50 +1,0 @@
-syntax = "proto3";
-
-package base;
-
-message Empty {
-}
-
-message ClientId {
-  string id = 1;
-}
-
-message SignalId {
-  string name = 1;
-  NameSpace namespace = 2;
-}
-
-message SignalInfo {
-  SignalId id = 1;
-  MetaData metaData = 2;
-}
-
-message MetaData {
-  string description = 4;
-  int32 max = 5;
-  int32 min = 6;
-  string unit = 7;
-  int32 size = 8;
-  bool isRaw = 9;
-  double factor = 10;
-  double offset = 11;
-}
-
-message NameSpace {
-  string name = 1;
-}
-
-message NetworkInfo {
-  NameSpace namespace = 1;
-  string type = 2;
-  string description = 3;
-}
-
-message FrameInfo {
-  SignalInfo signalInfo = 1;
-  repeated SignalInfo childInfo = 2;
-}
-
-message Frames {
-  repeated FrameInfo frame = 1;
-}
