--- rust/remotivelabs-broker/examples/pubsub/protos/network_api.proto
@@ -1,43 +1,0 @@
-syntax = "proto3";
-
-import "common.proto";
-
-package base;
-
-service NetworkService {
-  rpc SubscribeToSignals (SubscriberConfig) returns (stream Signals) {}
-  rpc PublishSignals (PublisherConfig) returns (Empty) {}
-  rpc ReadSignals (SignalIds) returns (Signals) {}
-}
-
-message SubscriberConfig {
-  ClientId clientId = 1;
-  SignalIds signals = 2;
-  bool onChange = 3;
-}
-
-message SignalIds {
-  repeated SignalId signalId = 1;
-}
-
-message Signals {
-  repeated Signal signal = 1;
-}
-
-message PublisherConfig {
-  Signals signals = 1;
-  ClientId clientId = 2;
-  int32 frequency = 3;
-}
-
-message Signal {
-  SignalId id = 1;
-  oneof payload {
-    int64 integer = 2;
-    double double = 3;
-    bool arbitration = 4;
-    bool empty = 6;
-  }
-  bytes raw = 5;
-  int64 timestamp = 7;
-}
