+++ rust/remotivelabs-broker/src/remotive_api/.gitignore
@@ -1,0 +1,1 @@
+base.rs
